% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_roclet.R
\name{info_roclet}
\alias{info_roclet}
\title{Roclet: check presence of \verb{@info} in documentation for exported functions}
\usage{
info_roclet()
}
\value{
A roclet to be used e.g. with \code{\link[roxygen2:roxygenize]{roxygen2::roxygenise()}}
}
\description{
This roclet allows you to add @info tag in roxygen2
}
\concept{roclets}
