% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_roclet.R
\name{information_roclet}
\alias{information_roclet}
\title{Roclet: check presence of \verb{@information} in documentation for exported functions}
\usage{
information_roclet()
}
\value{
A roclet to be used e.g. with \code{\link[roxygen2:roxygenize]{roxygen2::roxygenise()}}
}
\description{
This roclet allows you to add @information tag in roxygen2
}
\concept{roclets}
