% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_roclet.R
\name{roclet_output.roclet_information}
\alias{roclet_output.roclet_information}
\title{Roclet Output}
\usage{
\method{roclet_output}{roclet_information}(x, results, base_path, ...)
}
\arguments{
\item{x}{input}

\item{results}{result}

\item{base_path}{path}

\item{...}{additional arguments}
}
\description{
roclet output function
}
