% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_roclet.R
\name{roclet_process.roclet_info}
\alias{roclet_process.roclet_info}
\title{Roclet Process}
\usage{
\method{roclet_process}{roclet_info}(x, blocks, env, base_path)
}
\arguments{
\item{x}{input}

\item{blocks}{blocks}

\item{env}{environment}

\item{base_path}{path}
}
\description{
roclet process function
}
