% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_roclet.R
\name{roxy_tag_parse.roxy_tag_information}
\alias{roxy_tag_parse.roxy_tag_information}
\title{Roclet parse}
\usage{
\method{roxy_tag_parse}{roxy_tag_information}(x)
}
\arguments{
\item{x}{input}
}
\description{
Roclet parse
}
